[영환] product Detail 수정
 - profile header 를 component 로 수정
 - temper과 profile로 나눠서 component로 분리
